requirements.txt
absl-py==0.15.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.13.1
antlr4-python3-runtime==4.8
appdirs==1.4.4
argh==0.26.2
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==4.0.1
attrs==21.2.0
autopep8==1.6.0
av==8.0.3
backcall==0.2.0
backports.functools-lru-cache==1.6.4
bcrypt==4.1.2
beautifulsoup4==4.7.1
beautifultable==0.7.0
bleach==4.1.0
blinker==1.4
bokeh==2.4.2
boto3==1.21.22
botocore==1.24.22
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.4
celluloid==0.2.0
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.7
click
cloudpickle==2.0.0
cmake==3.28.1
colorama==0.4.4
configparser==5.1.0
cryptography==36.0.0
cycler==0.10.0
Cython==0.29.24
cytoolz==0.11.2
debugpy==1.4.1
decorator==5.1.0
decord==0.6.0
deepspeed==0.8.0
defusedxml==0.7.1
-f https://dl.fbaipublicfiles.com/detectron2/wheels/cu111/torch1.9/index.html
detectron2
dictdiffer==0.9.0
distributed==2021.11.2
docker==3.6.0
docker-pycreds==0.4.0
docopt==0.6.2
editdistance==0.6.0
ego4d==1.0
einops==0.3.0
entrypoints==0.3
envyaml==1.9.210927
fairscale==0.4.12
filelock==3.13.1
Flask==1.1.4
fonttools==4.25.0
frozenlist==1.2.0
fsspec
future==0.18.2
fvcore==0.1.5.post20211023
gevent==21.12.0
gevent-websocket==0.10.1
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.2
google-auth-oauthlib==0.4.6
gql==0.1.0
graphql-core==3.1.6
greenlet==1.1.2
grpcio==1.41.1
gulpio2==0.0.4
gunicorn==21.2.0
h5py==3.2.1
HeapDict==1.0.1
hjson==3.1.0
huggingface-hub==0.10.1
hydra-core==1.1.1
idna==2.10
imageio==2.10.3
importlib-metadata==4.8.1
importlib-resources==5.3.0
imutils==0.5.4
iopath==0.1.9
ipdb==0.13.9
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==1.0.0
joblib
jsonschema==4.1.2
jupyter-client==7.0.6
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.8.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.1
Lintel
lit==17.0.6
lmdb==1.2.1
locket==0.2.0
lxml==4.6.2
Mako==1.3.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.3
mephisto==0.4.2
mistune==0.8.4
mpmath==1.3.0
msgpack==1.0.2
multidict==5.2.0
munkres==1.1.4
mypy-extensions==0.4.3
natsort==7.1.1
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
ninja==1.11.1
nltk==3.5
notebook==6.4.5
numexpr==2.7.3
numpy==1.22.0
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
olefile==0.46
omegaconf==2.1.2
opencv-python==4.5.3.56
packaging==21.0
pandas==1.3.5
pandocfilters==1.5.0
parameterized==0.8.1
paramiko==3.4.0
parso==0.8.2
partd==1.2.0
patchify==0.2.3
pathspec==0.9.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
Pillow-SIMD==7.0.0.post3
pip==24.0
platformdirs==2.5.2
portalocker==2.3.2
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.21
protobuf==3.18.0
psutil==5.9.0
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyarrow==14.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.8.0
pycparser==2.20
pydantic==1.10.4
pyDeprecate==0.3.1
pydot==1.4.2
Pygments==2.10.0
PyJWT==1.7.1
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent==0.17.3
pysftp==0.2.9
PySocks==1.7.1
python-dateutil==2.7.3
pytorch-lightning==1.5.6
pytorchvideo==0.1.5
pytz==2021.3
pyu2f==0.1.5
PyWavelets==1.2.0
PyYAML==5.1
pyzmq==19.0.2
querystring-parser==1.2.4
regex==2020.10.15
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.2
sacremoses==0.0.47
safetensors==0.4.1
scikit-image==0.18.3
scikit-learn==1.1.3
scipy
Send2Trash==1.8.0
sentence-transformers==2.1.0
sentencepiece==0.1.96
sentry-sdk==1.5.0
setuptools==58.0.4
sh==1.14.2
shortuuid==1.0.8
simplejpeg==1.6.5
six==1.16.0
smmap==3.0.5
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.25
sqlparse==0.4.4
submitit==1.3.3
subprocess32==3.5.4
swifter==1.0.9
sympy==1.12
tabulate==0.8.9
tblib==1.7.0
tensorboard==2.7.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.11.2
timm==0.6.11
tokenizers
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
-f https://download.pytorch.org/whl/torch/
torch==1.9.1+cu111
torchmetrics==0.6.0
-f https://download.pytorch.org/whl/torchvision/
torchvision==0.10.1+cu111
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
transformers==4.16.1
triton==2.0.0
typing_extensions
urllib3==1.26.7
validators==0.12.6
wandb==0.12.1
warmup-scheduler==0.3
watchdog==0.10.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==1.0.1
wget==3.2
wheel==0.37.0
widgetsnbextension==3.5.1
xmltodict==0.12.0
yacs==0.1.8
yarl==1.7.2
zict==2.0.0
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
